+++ tiger-test-lib/src/test/java/de/gematik/test/tiger/lib/httpclient/TestHttpClientSteps.java
@@ -26,13 +26,15 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
 import java.util.concurrent.atomic.AtomicBoolean;
-import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
 import org.mockserver.client.MockServerClient;
 import org.mockserver.junit.jupiter.MockServerExtension;
 
@@ -62,6 +64,11 @@ public class TestHttpClientSteps {
       TigerDirector.getLibConfig().getHttpClientConfig().setActivateRbelWriter(true);
     }
     rbelValidatorGlueCode.tgrClearRecordedMessages();
+  }
+
+  @AfterEach
+  void clearDefaultHeaders() {
+    httpGlueCode.clearDefaultHeaders();
   }
 
   @AfterAll
@@ -105,7 +112,7 @@ public class TestHttpClientSteps {
         tigerResolvedString("!{rbel:currentRequestAsString('$.method')}"), "POST");
     tigerGlue.tgrAssertMatches(
         tigerResolvedString("!{rbel:currentRequestAsString('$.header.Content-Type')}"),
-        "application/json");
+        "application/json; charset=UTF-8");
     tigerGlue.tgrAssertMatches(
         tigerResolvedString("!{rbel:currentRequestAsString('$.body.object.field')}"), "value");
     tigerGlue.tgrAssertMatches(
@@ -154,7 +161,7 @@ public class TestHttpClientSteps {
         tigerResolvedString("!{rbel:currentRequestAsString('$.method')}"), "PUT");
     tigerGlue.tgrAssertMatches(
         tigerResolvedString("!{rbel:currentRequestAsString('$.header.Content-Type')}"),
-        "application/json");
+        "application/json; charset=UTF-8");
     tigerGlue.tgrAssertMatches(
         tigerResolvedString("!{rbel:currentRequestAsString('$.body.object.field')}"), "value");
     tigerGlue.tgrAssertMatches(
@@ -248,7 +255,7 @@ public class TestHttpClientSteps {
     rbelValidatorGlueCode.findLastRequestToPath(".*");
     tigerGlue.tgrAssertMatches(
         tigerResolvedString("!{rbel:currentRequestAsString('$.header.Content-Type')}"),
-        "text/plain.*");
+        "text/plain");
   }
 
   @Test
@@ -395,8 +402,48 @@ public class TestHttpClientSteps {
         "value with spaces");
   }
 
-  @SneakyThrows
+  @ParameterizedTest
+  @CsvSource(
+      textBlock =
+          """
+    {"hello": "world"}, application/json; charset=UTF-8
+    <this><is>xml</is></this>, application/xml; charset=ISO-8859-1
+    just some text, application/octet-stream; charset=ISO-8859-1
+    """)
+  void
+      sendRequestWithoutExplicitHeaders_shouldAutomaticallyAddContentTypeAndCharsetBasedOnContent( // NOSONAR
+      String bodyContent, String expectedContentTypeHeader) {
+    httpGlueCode.sendRequestWithBody(
+        Method.POST, createAddress("http://httpbin/headers"), bodyContent);
+    rbelValidatorGlueCode.findLastRequestToPath(".*");
+    tigerGlue.tgrAssertMatches(
+        tigerResolvedString("!{rbel:currentRequestAsString('$.header.Content-Type')}"),
+        expectedContentTypeHeader);
+  }
+
+  @ParameterizedTest
+  @CsvSource(
+      textBlock =
+          """
+        {"hello": "world"}, application/my-own-thing
+        <this><is>xml</is></this>, application/fancy-xml
+        just some text, application/text
+        {"hello": "world"}, application/json; charset=ISO-8859-1
+        <this><is>xml</is></this>, application/xml; charset=UTF-8
+        just some text, application/text; charset=UTF-8
+    """)
+  void sendRequestWithExplicitHeaders_shouldNotAddAnythingAutomatically( // NOSONAR
+      String bodyContent, String explicitSetContentTypeHeader) {
+    httpGlueCode.setDefaultHeader("Content-Type", explicitSetContentTypeHeader);
+    httpGlueCode.sendRequestWithBody(
+        Method.POST, createAddress("http://httpbin/headers"), bodyContent);
+    rbelValidatorGlueCode.findLastRequestToPath(".*");
+    tigerGlue.tgrAssertMatches(
+        tigerResolvedString("!{rbel:currentRequestAsString('$.header.Content-Type')}"),
+        explicitSetContentTypeHeader);
+  }
+
   private static URI createAddress(String urlString) {
-    return new URI(urlString);
+    return URI.create(urlString);
   }
 }